Add files via upload
Two straightforward scripts that enable snapping an object pivot to a vertex of a skinned mesh (although you can easily change the skinned mesh to a mesh filter). The first instantly snaps at the chosen vertex in runtime, and the second, is has the vertex position updated according to the value set on the coroutine and must be in a small defined distance from the vertex to allow the snapping. Very easy to change and manipulate the offsets.